requirements.txt
discord-py-slash-command
discord.py
requests
python-dotenv